backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
websockets
pinecone
groq
transformers
torch
bertopic
scikit-learn
nltk
